+++ AddOns/Baganator/ItemViewCommon/CurrencyBar.lua
@@ -109,7 +109,12 @@ local function ShowCurrencies(self, character)
       if strlenutf8(currencyText) > 5 then
         currencyText = AbbreviateNumbers(count)
       end
-      currencyText = currencyText .. " " .. CreateTextureMarkup(C_CurrencyInfo.GetCurrencyInfo(details.currencyID).iconFileID, 14, 14, 12, 12, 0.08, 0.96, 0.08, 0.96)
+      local currencyInfo = C_CurrencyInfo.GetCurrencyInfo(details.currencyID)
+      if currencyInfo then
+        currencyText = currencyText .. " " .. CreateTextureMarkup(currencyInfo.iconFileID, 14, 14, 12, 12, 0.08, 0.96, 0.08, 0.96)
+      else
+        currencyText = currencyText .. " ??"
+      end
       fontString:SetText(currencyText)
 
       if GameTooltip.SetCurrencyByID then