+++ AddOns/Baganator/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Baganator
 
-## [781](https://github.com/TheMouseNest/Baganator/tree/781) (2026-01-21)
-[Full Changelog](https://github.com/TheMouseNest/Baganator/compare/780...781) 
+## [782](https://github.com/TheMouseNest/Baganator/tree/782) (2026-01-25)
+[Full Changelog](https://github.com/TheMouseNest/Baganator/compare/781...782) 
 
-- Disable skins if the associated addon is incompatible with current wow version  
+- Fix error when previewing character with a removed currency from the game  
+- Add API to tell other addons which corner their widget has been assigned to